git: extract git helpers to src/server/git/
Move execGit, execGitSafe, getGitDiff, batchGitStatus, GitStatusResult,
PR-status caching, and deleteRemoteGoalBranches out of server.ts into
dedicated modules under src/server/git/.

Re-export GitStatusResult and the test-only hooks from server.ts so
existing imports from "./server.js" keep working.

Net: server.ts -291 lines (8433 → 8142). No behaviour change.
